backend/requirements.txt
https://download.pytorch.org/whl/cpu/torch-1.13.0%2Bcpu-cp310-cp310-linux_x86_64.whl
Cython==0.29.32
Flask==2.2.2
Flask_Cors==3.0.10
Pillow==9.3.0
gensim==4.2.0
gunicorn==20.1.0
nltk==3.7
newspaper3k==0.2.8
pandas==1.5.1
pymongo==4.3.2
scikit_learn==1.1.3
smart_open==5.2.1
spacy==3.4.3
textacy==0.12.0
transformers==4.24.0